+++ api/handlers/node_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"crypto"
-	"crypto/rand"
 	"crypto/rsa"
 	"crypto/sha256"
 	"encoding/hex"
@@ -19,8 +18,6 @@ import (
 	"go.uber.org/zap/zapcore"
 
 	"github.com/bloxapp/ssv/api"
-	p2pv1 "github.com/bloxapp/ssv/network/p2p"
-	"github.com/bloxapp/ssv/nodeprobe"
 	"github.com/bloxapp/ssv/utils/rsaencryption"
 )
 
@@ -33,7 +30,7 @@ func TestHandlers(t *testing.T) {
 	n := 4
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	node := createTestNode(t, priv, n, ctx, logger)
+	node := CreateTestNode(t, priv, n, ctx, logger)
 
 	t.Run("identity", func(t *testing.T) {
 		req, err := http.NewRequest("GET", "/v1/node/identity", nil)
@@ -128,31 +125,4 @@ func TestHandlers(t *testing.T) {
 		err = json.Unmarshal(rr.Body.Bytes(), &topics)
 		require.NoError(t, err)
 	})
-}
-
-func createTestNode(t *testing.T, priv *rsa.PrivateKey, n int, ctx context.Context, logger *zap.Logger) *Node {
-	pks := []string{"b768cdc2b2e0a859052bf04d1cd66383c96d95096a5287d08151494ce709556ba39c1300fbb902a0e2ebb7c31dc4e400",
-		"824b9024767a01b56790a72afb5f18bb0f97d5bddb946a7bd8dd35cc607c35a4d76be21f24f484d0d478b99dc63ed170"}
-	ln, routers, err := api.CreateNetworkAndSubscribe(t, ctx, p2pv1.LocalNetOptions{
-		Nodes:        n,
-		MinConnected: n/2 - 1,
-		UseDiscv5:    false,
-	}, pks...)
-	require.NoError(t, err)
-	require.NotNil(t, routers)
-	require.NotNil(t, ln)
-
-	node := &api.Node{}
-	node.HealthyMock.Store(nil)
-	nodeProber := nodeprobe.NewProber(zap.L(), nil, map[string]nodeprobe.Node{"consensus client": node, "execution client": node, "event syncer": node})
-	return &Node{
-		ListenAddresses: []string{fmt.Sprintf("tcp://%s:%d", "localhost", 3030), fmt.Sprintf("udp://%s:%d", "localhost", 3030)},
-		PeersIndex:      ln.Nodes[0].(p2pv1.PeersIndexProvider).PeersIndex(),
-		Network:         ln.Nodes[0].(p2pv1.HostProvider).Host().Network(),
-		TopicIndex:      ln.Nodes[0].(TopicIndex),
-		NodeProber:      nodeProber,
-		Signer: func(data []byte) ([]byte, error) {
-			return rsa.SignPKCS1v15(rand.Reader, priv, crypto.SHA256, data[:])
-		},
-	}
 }
