+++ api/handlers/test_utils.go
@@ -1,8 +1,12 @@
-package api
+package handlers
 
 import (
 	"context"
+	"crypto"
+	"crypto/rand"
+	"crypto/rsa"
 	"encoding/hex"
+	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -20,10 +24,38 @@ import (
 	"github.com/bloxapp/ssv/message/validation"
 	"github.com/bloxapp/ssv/network"
 	p2pv1 "github.com/bloxapp/ssv/network/p2p"
+	"github.com/bloxapp/ssv/nodeprobe"
 	"github.com/bloxapp/ssv/protocol/v2/ssv/queue"
 )
 
-func CreateNetworkAndSubscribe(t *testing.T, ctx context.Context, options p2pv1.LocalNetOptions, pks ...string) (*p2pv1.LocalNet, []*dummyRouter, error) {
+func CreateTestNode(t *testing.T, priv *rsa.PrivateKey, n int, ctx context.Context, logger *zap.Logger) *Node {
+	pks := []string{"b768cdc2b2e0a859052bf04d1cd66383c96d95096a5287d08151494ce709556ba39c1300fbb902a0e2ebb7c31dc4e400",
+		"824b9024767a01b56790a72afb5f18bb0f97d5bddb946a7bd8dd35cc607c35a4d76be21f24f484d0d478b99dc63ed170"}
+	ln, routers, err := createNetworkAndSubscribe(t, ctx, p2pv1.LocalNetOptions{
+		Nodes:        n,
+		MinConnected: n/2 - 1,
+		UseDiscv5:    false,
+	}, pks...)
+	require.NoError(t, err)
+	require.NotNil(t, routers)
+	require.NotNil(t, ln)
+
+	node := &NodeMock{}
+	node.HealthyMock.Store(nil)
+	nodeProber := nodeprobe.NewProber(zap.L(), nil, map[string]nodeprobe.Node{"consensus client": node, "execution client": node, "event syncer": node})
+	return &Node{
+		ListenAddresses: []string{fmt.Sprintf("tcp://%s:%d", "localhost", 3030), fmt.Sprintf("udp://%s:%d", "localhost", 3030)},
+		PeersIndex:      ln.Nodes[0].(p2pv1.PeersIndexProvider).PeersIndex(),
+		Network:         ln.Nodes[0].(p2pv1.HostProvider).Host().Network(),
+		TopicIndex:      ln.Nodes[0].(TopicIndex),
+		NodeProber:      nodeProber,
+		Signer: func(data []byte) ([]byte, error) {
+			return rsa.SignPKCS1v15(rand.Reader, priv, crypto.SHA256, data[:])
+		},
+	}
+}
+
+func createNetworkAndSubscribe(t *testing.T, ctx context.Context, options p2pv1.LocalNetOptions, pks ...string) (*p2pv1.LocalNet, []*dummyRouter, error) {
 	logger, err := zap.NewDevelopment()
 	require.NoError(t, err)
 	ln, err := p2pv1.CreateAndStartLocalNet(ctx, logger.Named("createNetworkAndSubscribe"), options)
@@ -104,11 +136,11 @@ func (r *dummyRouter) Route(_ context.Context, _ *queue.DecodedSSVMessage) {
 	atomic.AddUint64(&r.count, 1)
 }
 
-type Node struct {
+type NodeMock struct {
 	HealthyMock atomic.Pointer[error]
 }
 
-func (sc *Node) Healthy(context.Context) error {
+func (sc *NodeMock) Healthy(context.Context) error {
 	err := sc.HealthyMock.Load()
 	if err != nil {
 		return *err
